v1.4.67: 修复装备孕养等级问题，优化任务阁和宗门消息
调整:
- 任务阁每年刷新任务数量从1-7个调整为1-5个
- 移除任务卡片上的奖励预览显示
- 移除AI宗门之间的外交消息
- 移除发现洞府的消息通知
- 移除AI宗门探索洞府成功的消息通知

修复:
- 修复装备孕养等级无法提升的问题
  - nurtureProgress字段类型从Int改为Double
  - 修复每tick经验增量被截断为0的问题

系统:
- 数据库版本升级到50

+++ android/app/src/main/java/com/xianxia/sect/core/engine/AISectDiscipleManager.kt
@@ -308,7 +308,7 @@ object AISectDiscipleManager {
                 equipmentId = equipmentId,
                 rarity = rarity,
                 nurtureLevel = 0,
-                nurtureProgress = 0
+                nurtureProgress = 0.0
             )
         } else {
             currentNurture
@@ -321,14 +321,14 @@ object AISectDiscipleManager {
         val monthlyGain = autoExpGain * SECONDS_PER_MONTH
         val expRequired = EquipmentNurtureSystem.getExpRequiredForLevelUp(nurture.nurtureLevel, rarity)
         
-        val newProgress = nurture.nurtureProgress + monthlyGain.toInt()
+        val newProgress = nurture.nurtureProgress + monthlyGain
         
         return if (newProgress >= expRequired) {
             val newLevel = (nurture.nurtureLevel + 1).coerceAtMost(maxLevel)
-            val remainingExp = (newProgress - expRequired).toInt()
+            val remainingExp = newProgress - expRequired
             nurtureSetter(disciple, nurture.copy(
                 nurtureLevel = newLevel,
-                nurtureProgress = if (newLevel >= maxLevel) 0 else remainingExp
+                nurtureProgress = if (newLevel >= maxLevel) 0.0 else remainingExp
             ))
         } else {
             nurtureSetter(disciple, nurture.copy(nurtureProgress = newProgress))
